+++ Samples~/Scripts/IssueAndTesting/Issue/Issue157/UltEventBase.cs
@@ -1,0 +1,10 @@
+using System;
+using UnityEngine.Events;
+
+namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue157
+{
+    [Serializable]
+    public class UltEventBase : UnityEvent
+    {
+    }
+}